+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilterTable.java
@@ -917,7 +917,7 @@ public final class FilterTable<R, C> extends JTable {
 	}
 
 	private void configureFilterConditionsPanel(ColumnConditionsPanel<C> filterConditionsPanel) {
-		filterConditionsPanel.panels().forEach(this::configureFilterPanel);
+		filterConditionsPanel.get().forEach(this::configureFilterPanel);
 	}
 
 	private void configureFilterPanel(C identifier, ConditionPanel<?> filterPanel) {